Reenable visualization tests
